Move preferred locale APIs to RevenueCat from RevenueCatUI
Now these APIs do not require importing RevenueCatUI
